* graphics/RenderSurface.* : StretchBlit function sort of implemented 	* graphics/SoftRenderSurface.* : Point scaling for StretchBlit 	* graphics/scalers/PointScaler.h : A simple point scaler. Work in 	  Progress 	* gumps/ScalerGump.* : Scaler gump. 	* kernel/GUIApp.* : Basic scaler support. Set 'scalex' and 'scaley' to 	  the desired scaling factor. Only integer factors supported. Note 	  it only works IF the backbuffer resolution can be evenly divided by 	  scale factor. So you can't do width=640 with scalex=3 (640 can't be 	  divided by 3). If a scale factor number is negetive, then it is taken 	  to mean an absolute size (positive of course). Same with factors over 	  100. Non integer scaling factors WILL be supported.

+++ graphics/RenderSurface.cpp
@@ -29,24 +29,24 @@ Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
 
 uint32	RenderSurface::s_bpp;
 
-uint8	RenderSurface::r_loss = 0;
-uint8	RenderSurface::r_loss16 = 0;
-uint8	RenderSurface::r_shift = 0;
+uint32	RenderSurface::r_loss = 0;
+uint32	RenderSurface::r_loss16 = 0;
+uint32	RenderSurface::r_shift = 0;
 uint32	RenderSurface::r_mask = 0;
 
-uint8	RenderSurface::g_loss = 0;
-uint8	RenderSurface::g_loss16 = 0;
-uint8	RenderSurface::g_shift = 0;
+uint32	RenderSurface::g_loss = 0;
+uint32	RenderSurface::g_loss16 = 0;
+uint32	RenderSurface::g_shift = 0;
 uint32	RenderSurface::g_mask = 0;
 
-uint8	RenderSurface::b_loss = 0;
-uint8	RenderSurface::b_loss16 = 0;
-uint8	RenderSurface::b_shift = 0;
+uint32	RenderSurface::b_loss = 0;
+uint32	RenderSurface::b_loss16 = 0;
+uint32	RenderSurface::b_shift = 0;
 uint32	RenderSurface::b_mask = 0;
 
-uint8	RenderSurface::a_loss = 0;
-uint8	RenderSurface::a_loss16 = 0;
-uint8	RenderSurface::a_shift = 0;
+uint32	RenderSurface::a_loss = 0;
+uint32	RenderSurface::a_loss16 = 0;
+uint32	RenderSurface::a_shift = 0;
 uint32	RenderSurface::a_mask = 0;
